refactorizacion de la estructura de directorios
