+++ README.md
@@ -9,29 +9,29 @@
 
 <div style="display: flex; margin-bottom: 50px" >
 
-<img src="img-readme/W - Inicio.png" alt="wireframe inicio">
+<img src="src/assets/img-readme/W - Inicio.png" alt="wireframe inicio">
 
-<img src="img-readme/W - Clima.png" style="margin: 0 50px" alt="wireframe clima">
+<img src="src/assets/img-readme/W - Clima.png" style="margin: 0 50px" alt="wireframe clima">
 
-<img src="img-readme/Clima - Mapa.png" alt="wireframe mapa">
-
-</div>
-
-<div style="display: flex; margin-bottom: 50px" >
-
-<img src="img-readme/W - Quienes somos - Nosotros.png" alt="wireframe nosotros">
-
-<img src="img-readme/W - Quienes somos - Solución.png" style="margin: 0 50px" alt="wireframe solucion">
-
-<img src="img-readme/W - Quienes somos - Diseño.png" alt="wireframe diseño">
+<img src="src/assets/img-readme/Clima - Mapa.png" alt="wireframe mapa">
 
 </div>
 
 <div style="display: flex; margin-bottom: 50px" >
 
-<img src="img-readme/W - Ciudades.png" alt="wireframe ciudades">
+<img src="src/assets/img-readme/W - Quienes somos - Nosotros.png" alt="wireframe nosotros">
 
-<img src="img-readme/Ciudades - modal.png" style="margin: 0 50px" alt="wireframe modal">
+<img src="src/assets/img-readme/W - Quienes somos - Solución.png" style="margin: 0 50px" alt="wireframe solucion">
+
+<img src="src/assets/img-readme/W - Quienes somos - Diseño.png" alt="wireframe diseño">
+
+</div>
+
+<div style="display: flex; margin-bottom: 50px" >
+
+<img src="src/assets/img-readme/W - Ciudades.png" alt="wireframe ciudades">
+
+<img src="src/assets/img-readme/Ciudades - modal.png" style="margin: 0 50px" alt="wireframe modal">
 
 </div>
 
@@ -39,20 +39,20 @@
 
 <div style="display: flex; margin-bottom: 50px" >
 
-<img src="img-readme/M - Inicio.png" alt="mockups inicio">
+<img src="src/assets/img-readme/M - Inicio.png" alt="mockups inicio">
 
 </div>
 
 <div style="display: flex; margin-bottom: 50px" >
 
-<img src="img-readme/M - Quienes somos - Nosotros.png" alt="mockups nosotros">
+<img src="src/assets/img-readme/M - Quienes somos - Nosotros.png" alt="mockups nosotros">
 
-<img src="img-readme/M - Quienes somos - Solución.png" style="margin: 0 50px" alt="mockups solucion">
+<img src="src/assets/img-readme/M - Quienes somos - Solución.png" style="margin: 0 50px" alt="mockups solucion">
 
-<img src="img-readme/M - Quienes somos - Diseño.png" alt="mockups diseño">
+<img src="src/assets/img-readme/M - Quienes somos - Diseño.png" alt="mockups diseño">
 
 </div>
 
 ## Persona Canva
 
-<img src="img-readme/Persona Canva.png" alt="Persona canva">
+<img src="src/assets/img-readme/Persona Canva.png" alt="Persona canva">
