+++ example.env
@@ -1,3 +1,2 @@
-API_KEY = "insert-your-key";
-OPENWEATHERAPI_KEY= "insert-your-openweatherapi-key";
+API_KEY = "insert-your-openweatherapi-key";
 FIREBASE_KEY = "insert-your-firebase-key";
